Separating directive into standalone files 1) Created new directives for each mention inside directive.js 2) Removed directive.js
